+++ install/js/mb/admin/kit/src/index.js
@@ -4,11 +4,12 @@ import * as Form from './form-save';
 import * as Dependencies from './dependencies';
 import * as Visibility from './visibility';
 import * as OptionsPage from './options-page';
-
+import * as GridCollapsible from './grid-collapsible';
 
 export {
     Form,
     Dependencies,
     Visibility,
-    OptionsPage
+    OptionsPage,
+    GridCollapsible,
 };